+++ apps/client/src/components/tiercard/index.tsx
@@ -12,7 +12,7 @@ type TierCardComponentProps = {
   handleClick: (planName: string) => void
 }
 
-export function TierCardComponent({ disabled, planName, price, grantedCredits, features, handleClick }: TierCardComponentProps) {
+export function TierCardComponent({ disabled, planName, price, grantedCredits, features }: TierCardComponentProps) {
   return (
     <div className="grid w-full items-start gap-10 rounded-lg border p-10 md:grid-cols-[1fr_200px]">
       <div className="grid gap-6">
@@ -28,13 +28,13 @@ export function TierCardComponent({ disabled, planName, price, grantedCredits, f
             ))
           }
           <li className="flex items-center text-slate-600">
-            <Check className="mr-2 h-4 w-4" /> {Number(grantedCredits).toLocaleString()} API Requests
+            <Check className="mr-2 h-4 w-4" /> $ {Number(grantedCredits).toLocaleString()} API Requests
           </li>
         </ul>
       </div>
       <div className="flex flex-col gap-4 text-center">
         <div>
-          <h4 className="text-4xl font-bold">₹{price}</h4>
+          <h4 className="text-4xl font-bold">${price}</h4>
           <p className="text-sm font-medium text-muted-foreground">
             Billed Monthly
           </p>