+++ apps/client/src/events/eventUnion.ts
@@ -1,4 +1,0 @@
-export type EventUnion = {
-  OrganizationChangeEvent: () => Promise<void>
-  SearchEvent: (searchKeyword: string) => Promise<void>
-}