+++ apps/server/src/auth/credential.guard.ts
@@ -2,7 +2,7 @@ import { CanActivate, ExecutionContext, ForbiddenException, Injectable } from "@
 import { statusMessages } from "src/utils/constants/status-messages"
 import { EventEmitter2 } from "@nestjs/event-emitter"
 import { EventsUnion } from "src/core/events/events.union"
-import { subscriptionConfig } from "src/core/api/subscription/subscription.config"
+import { Products, subscriptionConfig } from "src/core/api/subscription/subscription.config"
 import { ModRequest } from "./types/mod-request.interface"
 import { User } from "src/core/api/user/schemas/user.schema"
 import { Organization } from "src/core/api/organization/schemas/organization.schema"
@@ -40,19 +40,19 @@ export class CredentialGuard implements CanActivate {
       const orgId = String(organization.id)
       const currentDate = new Date()
       const expiryDate = subscription.expiresAt
-      const { method, url: apiUri } = request
 
       if (currentDate > expiryDate) {
         throw new ForbiddenException(statusMessages.subscriptionExpired)
       }
 
-      const creditRequired = 1
+      const product = request.url.split("/")[3]
+      const { responseDelay, estimatedRequestCost } = subscriptionConfig.find((sub) => sub.planName === subscription.selectedPlan)
+      const creditRequired = estimatedRequestCost[product as Products]
 
       if (creditRequired > subscription.remainingCredits) {
         throw new ForbiddenException(statusMessages.subscriptionLimitReached)
       }
 
-      const responseDelay = subscriptionConfig.find((sub) => sub.planName === subscription.selectedPlan).responseDelay
       await new Promise(resolve => setTimeout(resolve, responseDelay))
       subscription.remainingCredits -= creditRequired
       await subscription.save()
@@ -66,6 +66,7 @@ export class CredentialGuard implements CanActivate {
       request.user = { userId, orgId }
 
       if (usageInsights) {
+        const { method, url: apiUri } = request
         this.eventEmitter.emit(EventsUnion.CreateInsights, { userId, method, apiUri })
       }