+++ apps/server/src/core/api/user/commands/impl/update-selected-org.command.ts
@@ -1,6 +1,0 @@
-export class UpdateSelectedOrgCommand {
-  constructor(
-    public readonly id: string,
-    public readonly orgId: string
-  ) { }
-}