+++ apps/server/src/core/api/user/commands/impl/update-usage-insights.command.ts
@@ -1,6 +1,0 @@
-export class UpdateUsageInsightsSettingsCommand {
-  constructor(
-    public readonly id: string,
-    public readonly usageInsights: boolean
-  ) { }
-}